requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
bcrypt==5.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0